feat: переведена на ssr страница edit/create
